requirements.txt
numpy==1.24.1
Pillow==9.4.0
PyAudio==0.2.13
pyserial==3.5
pystray==0.19.4
pynput~=1.7.6
PyAutoGUI~=0.9.53
PyGetWindow~=0.0.9
screeninfo~=0.8.1